move chunking tools into utils._chunking
